Addressing review comments: - Refactor and use ThreadSafeMutableRoaringBitmap for thread safety - Other minor fixes
