+++ pinot-core/src/main/java/org/apache/pinot/core/realtime/impl/invertedindex/RealtimeInvertedIndexReader.java
@@ -21,6 +21,7 @@ package org.apache.pinot.core.realtime.impl.invertedindex;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
+import org.apache.pinot.core.realtime.impl.ThreadSafeMutableRoaringBitmap;
 import org.apache.pinot.core.segment.index.readers.InvertedIndexReader;
 import org.roaringbitmap.buffer.MutableRoaringBitmap;
 
@@ -80,29 +81,5 @@ public class RealtimeInvertedIndexReader implements InvertedIndexReader<MutableR
 
   @Override
   public void close() {
-  }
-
-  /**
-   * Helper wrapper class for {@link MutableRoaringBitmap} to make it thread-safe.
-   */
-  private static class ThreadSafeMutableRoaringBitmap {
-    private MutableRoaringBitmap _mutableRoaringBitmap;
-
-    public ThreadSafeMutableRoaringBitmap(int firstDocId) {
-      _mutableRoaringBitmap = new MutableRoaringBitmap();
-      _mutableRoaringBitmap.add(firstDocId);
-    }
-
-    public void checkAndAdd(int docId) {
-      if (!_mutableRoaringBitmap.contains(docId)) {
-        synchronized (this) {
-          _mutableRoaringBitmap.add(docId);
-        }
-      }
-    }
-
-    public synchronized MutableRoaringBitmap getMutableRoaringBitmap() {
-      return _mutableRoaringBitmap.clone();
-    }
   }
 }
